requirements.txt
aiogram>=3.4.1
aiohttp>=3.9.1
playwright>=1.44.0
python-dotenv>=1.0.0
greenlet>=3.1.0
